refactor: code cleanup
