+++ src/internal/database/database.go
@@ -1,4 +1,4 @@
-package internal
+package database
 
 import (
 	"log"
